+++ src/main/java/com/driver/services/UserService.java
@@ -5,6 +5,7 @@ import com.driver.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Service
@@ -15,13 +16,17 @@ public class UserService {
     BlogService blogService;
 
     public User createUser(String username, String password){
-        User user = new User(username,password);
+        User user = new User();
+        user.setUsername(username);
+        user.setPassword(password);
+        user.setBlogList(new ArrayList<>());
         userRepository3.save(user);
         return user;
     }
 
     public void deleteUser(int userId){
         User user = userRepository3.findById(userId).orElse(null);
+        if(user == null) return ;
         List<Blog> blogList = user.getBlogList();
         for(Blog blog : blogList){
             blogService.deleteBlog(blog.getId());
@@ -31,6 +36,7 @@ public class UserService {
 
     public User updateUser(Integer id, String password){
         User user = userRepository3.findById(id).orElse(null);
+        if(user == null) return null;
         userRepository3.deleteById(id);
         user.setPassword(password);
         userRepository3.save(user);
